+++ CHANGELOG.md
@@ -16,6 +16,55 @@
 ### Security
 -->
 
+
+## [v3.13.1] - 2020-12-06
+
+### Changed
+
+- Minor UI fixes
+
+### Removed
+
+- Alert from CSV Output
+
+### Fixed
+
+- macOS Big Sur menu icon dark mode color
+
+
+## [v3.13.0] - 2020-09-23
+
+### Added
+
+- ALERT component to show messages in the app
+- More options for the HTTP component
+
+### Removed
+
+- FOCUS_WINDOW component
+
+## [v3.12.0] - 2020-09-19
+
+### Added
+
+- FOCUS_WINDOW component
+- Option to choose a delimiter character for the CSV_LOOKUP component
+- Hide option to the tray icon (Linux only)
+- Variables injection support for NUMBER and TEXT components
+
+### Changed
+
+- RUN, HTTP and CSV_LOOKUP Components are now executed synchronously
+- You can now use run, http and csv_lookup variables
+- Code Signing certificate
+- Increased RUN Component output buffer
+
+### Fixed
+
+- Keyboard emulation issue with special characters (Linux only)
+- Empty output when using CSV_LOOKUP
+- Fix "enableTray" error
+
 ## [v3.11.2] - 2020-07-29
 
 ### Fixed
@@ -329,6 +378,9 @@
 
 - Included the Apple Bonjour installer
 
+[v3.13.1]: https://github.com/fttx/barcode-to-pc-server/compare/v3.13.0...v3.13.1
+[v3.13.0]: https://github.com/fttx/barcode-to-pc-server/compare/v3.12.0...v3.13.0
+[v3.12.0]: https://github.com/fttx/barcode-to-pc-server/compare/v3.11.2...v3.12.0
 [v3.11.2]: https://github.com/fttx/barcode-to-pc-server/compare/v3.11.1...v3.11.2
 [v3.11.1]: https://github.com/fttx/barcode-to-pc-server/compare/v3.11.0...v3.11.1
 [v3.11.0]: https://github.com/fttx/barcode-to-pc-server/compare/v3.10.0...v3.11.0